app/requirements.txt
redis==2.10.6
Flask==0.12.2
psycopg2==2.6.1
Flask-SQLAlchemy===2.1
Flask-Migrate==1.8.0
grpcio==1.6.3
mock==2.0.0
protobuf==3.4.0
tensorflow==1.3.0
tensorflow-serving-api==1.3.0
virtualenv==15.1.0
